+++ practical-work-1/src/Utilitiles.cpp
@@ -16,18 +16,18 @@ bool fileExists(const string &fileName) {
 }
 
 Point3D normalizeVector(Point3D normal) {
-	double powX = pow(normal.getX(), 2);
-	double powY = pow(normal.getY(), 2);
-	double powZ = pow(normal.getZ(), 2);
+	float powX = pow(normal.getX(), 2);
+	float powY = pow(normal.getY(), 2);
+	float powZ = pow(normal.getZ(), 2);
 
-	double length = sqrt(powX + powY + powZ);
+	float length = sqrt(powX + powY + powZ);
 
 	return normal /= length;
 }
 
 // newell's method
 Point3D calculateSurfaceNormal(vector<Point3D> polygonVertexes) {
-	double x = 0, y = 0, z = 0;
+	float x = 0, y = 0, z = 0;
 
 	for (unsigned int i = 0; i < polygonVertexes.size(); i++) {
 		Point3D current = polygonVertexes[i];
